+++ biocuda/__init__.py
@@ -6,6 +6,8 @@ from .formulas import FormulaEngine
 from .planner import OptimizationPlanner
 from .specs import GPUSpec, GPU_DATABASE
 from .workloads import Workload, WorkloadType
+from . import kernels
+from . import tier_m
 
 __all__ = [
     "EXP3Tuner",
@@ -18,6 +20,8 @@ __all__ = [
     "WorkloadType",
     "autotune_plan",
     "detect_current_gpu",
+    "kernels",
+    "tier_m",
 ]
 
-__version__ = "39.1.0"
+__version__ = "39.3.0"